Module forest_fil_types::verifier

Functions

generate_winning_post_sector_challenge

Generates sector challenge indexes for use in winning PoSt verification.

verify_winning_post

Functionality for verification of seal, winning PoSt and window PoSt proofs. Proof verification will be full validation by default. Verifies winning proof of spacetime. These proofs are generated by the miners that are elected to mine a new block to verify a sector. A failed winning proof leads to a miner being slashed.