Trait forest_fil_types::verifier::ProofVerifier

pub trait ProofVerifierwhere
    Self: Sized + Send + Sync + 'static,{
    fn verify_winning_post(
        Randomness: Randomness,
        proofs: &[PoStProof],
        challenge_sectors: &[SectorInfo],
        prover: u64
    ) -> Result<(), Error> { ... }
    fn generate_winning_post_sector_challenge(
        proof: RegisteredPoStProof,
        prover_id: u64,
        Randomness: Randomness,
        eligible_sector_count: u64
    ) -> Result<Vec<u64>, Error> { ... }
}

Functionality for verification of seal, winning PoSt and window PoSt proofs. Proof verification will be full validation by default.

Provided Methods

fn verify_winning_post(
Randomness: Randomness,
proofs: &[PoStProof],
challenge_sectors: &[SectorInfo],
prover: u64
) -> Result<(), Error>

Verifies winning proof of spacetime. These proofs are generated by the miners that are elected to mine a new block to verify a sector. A failed winning proof leads to a miner being slashed.

fn generate_winning_post_sector_challenge(
proof: RegisteredPoStProof,
prover_id: u64,
Randomness: Randomness,
eligible_sector_count: u64
) -> Result<Vec<u64>, Error>

Generates sector challenge indexes for use in winning PoSt verification.